+++ RecordParser/Extensions/FileReader/VariableLengthReaderRawExtensions.cs
@@ -1,5 +1,4 @@
 ﻿using RecordParser.Engines.Reader;
-using RecordParser.Parsers;
 using System;
 using System.Buffers;
 using System.Collections.Generic;
@@ -10,34 +9,25 @@ using static RecordParser.Extensions.FileReader.ReaderCommon;
 
 namespace RecordParser.Extensions.FileReader
 {
-    public class VariableLengthReaderOptions
-    {
-        public bool hasHeader;
-        public bool parallelProcessing;
-        public bool containsQuotedFields;
-    }
-
     public delegate string StringFactory(ReadOnlySpan<char> text);
+    internal delegate void Get(ref TextFindHelper finder, string[] inst, StringFactory cache);
 
     public class VariableLengthReaderRawOptions
     {
         public bool hasHeader;
         public bool parallelProcessing;
         public bool containsQuotedFields;
+        public bool trim;
 
         public int columnCount;
         public string separator;
         public Func<StringFactory> StringFactory;
     }
 
-    public static partial class Exasd
+    public static class VariableLengthReaderRawExtensions
     {
-        private static int length = (int)Math.Pow(2, 23);
-
-        delegate void Get(ref TextFindHelper finder, string[] inst, StringFactory cache);
-        private static Get BuildRaw(int collumnCount, bool hasTransform)
+        private static Get BuildRaw(int collumnCount, bool hasTransform, bool trim)
         {
-            // parameters
             var configParameter = Expression.Parameter(typeof(TextFindHelper).MakeByRefType(), "config");
             var instanceVariable = Expression.Parameter(typeof(string[]), "inst");
             var cacheParameter = Expression.Parameter(typeof(StringFactory), "cache");
@@ -47,8 +37,9 @@ namespace RecordParser.Extensions.FileReader
             {
                 var arrayAccessExpr = Expression.ArrayAccess(instanceVariable, Expression.Constant(i));
                 var getValue = (Expression)Expression.Call(configParameter, nameof(TextFindHelper.GetValue), Type.EmptyTypes, Expression.Constant(i));
-                getValue = Expression.Call(typeof(MemoryExtensions), "Trim", Type.EmptyTypes, getValue);
 
+                if (trim)
+                    getValue = Expression.Call(typeof(MemoryExtensions), "Trim", Type.EmptyTypes, getValue);
 
                 if (hasTransform)
                 {
@@ -68,14 +59,13 @@ namespace RecordParser.Extensions.FileReader
             return final.Compile();
         }
 
-        // ???
         public static IEnumerable<T> GetRecordsRaw<T>(this TextReader stream, VariableLengthReaderRawOptions options, Func<Func<int, string>, T> reader)
         {
-            var get = BuildRaw(options.columnCount, options.StringFactory != null);
+            var get = BuildRaw(options.columnCount, options.StringFactory != null, options.trim);
 
             Func<IFL> func = options.containsQuotedFields
-                           ? () => new QuotedRow(stream, length, options.separator)
-                           : () => new RowByLine(stream, length);
+                           ? () => new RowByQuote(stream, Length, options.separator)
+                           : () => new RowByLine(stream, Length);
 
             return options.parallelProcessing
                     ? GetParallel()
@@ -136,7 +126,6 @@ namespace RecordParser.Extensions.FileReader
                         var r = funcs[i % maxParallelism];
                         lock (r.lockObj)
                         {
-
                             get(ref finder, r.buffer, r.stringCache);
 
                             return reader(r.getField);
@@ -148,19 +137,6 @@ namespace RecordParser.Extensions.FileReader
                     }
                 }
             }
-        }
-
-        public static IEnumerable<T> GetRecords<T>(this IVariableLengthReader<T> reader, TextReader stream, VariableLengthReaderOptions options)
-        {
-            Func<IFL> func = options.containsQuotedFields
-                            ? () => new QuotedRow(stream, length, reader.separator)
-                            : () => new RowByLine(stream, length);
-
-            Func<ReadOnlyMemory<char>, int, T> parser = (memory, i) => reader.Parse(memory.Span);
-
-            return options.parallelProcessing
-                    ? GetRecordsParallel(parser, func, options.hasHeader)
-                    : GetRecordsSequential(parser, func, options.hasHeader);
         }
     }
 }
